auth/requirements.txt
fastapi
uvicorn
pymongo
python-jose
python-dotenv
